Fix navigation bug, move ColorPalete component, and revise HomeScreen
