+++ src/components/ColorPalete/index.jsx
