+++ src/pages/HomeScreen/index.jsx
@@ -9,6 +9,15 @@ import { STORAGE_LOGIN_INFO } from 'shared/constants/storage';
 import storage from 'shared/storage';
 import logger from 'shared/utils/logger';
 
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    backgroundColor: '#fff',
+    alignItems: 'center',
+    justifyContent: 'center',
+  },
+});
+
 const HomeBox = ({ text, bg, onPress }) => (
   <TouchableOpacity onPress={onPress}>
     <Center w='64' h='64' bg={bg} rounded='md' shadow={3} _text={{ color: 'white' }}>
